pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.39",
    "streamlit-authenticator>=0.4.2",
    "streamlit>=1.43.2",
]
